Revert "Create /audio page (#279)"
This reverts commit 97e8e7fad4e829118671687dd07a5823de78f9af.
