+++ src/containers/artist-dashboard-page/Tiles.module.css
@@ -127,6 +127,86 @@
   box-shadow: 0px 5px 15px -2px rgba(0, 0, 0, 0.25);
 }
 
+/* Explainer Tile */
+.explainerTile {
+  height: 331px;
+  background-color: var(--white);
+  display: inline-flex;
+  flex-wrap: wrap;
+  flex-direction: row;
+  justify-content: flex-start;
+  align-items: flex-start;
+  overflow: hidden;
+  text-align: left;
+}
+
+.whatIsAudioContainer {
+  flex: 1;
+  display: inline-flex;
+  flex-direction: column;
+  align-items: flex-start;
+  padding: 44px 0px;
+  margin-bottom: 74px;
+}
+
+@media screen and (max-width: 1024px) {
+  .whatIsAudioContainer {
+    padding-right: 32px;
+  }
+}
+
+.whatIsAudio {
+  font-weight: 800;
+  font-size: 40px;
+  line-height: 49px;
+  background: var(--page-header-gradient);
+  -webkit-background-clip: text;
+  -webkit-text-fill-color: transparent;
+  margin: 0px 0px 2px 0px;
+}
+
+.whyCare {
+  font-weight: var(--font-bold);
+  font-size: var(--font-m);
+  line-height: 20px;
+  text-transform: uppercase;
+  color: var(--neutral-light-4);
+  margin-bottom: 18px;
+  text-align: left;
+}
+
+.description {
+  font-weight: var(--font-demi-bold);
+  font-size: 20px;
+  line-height: 160%;
+  color: var(--neutral);
+  margin-bottom: 16px;
+  max-width: 640px;
+}
+
+.learnMore {
+  font-weight: var(--font-bold);
+  font-size: 20px;
+  line-height: 17px;
+  text-align: right;
+  letter-spacing: 0.5px;
+  color: var(--secondary);
+  margin: 10px 0px 8px;
+  cursor: pointer;
+}
+
+.learnMore:hover {
+  text-decoration: underline;
+}
+
+.confused {
+  font-weight: var(--font-demi-bold);
+  font-size: var(--font-m);
+  line-height: 24px;
+  color: var(--neutral);
+  margin-bottom: 28px;
+}
+
 .levels {
   display: inline-flex;
   flex-direction: column;
