+++ src/containers/artist-dashboard-page/WalletModal.module.css
