+++ src/containers/artist-dashboard-page/WalletModal.tsx
