+++ src/containers/artist-dashboard-page/components/ClaimSuccessBody.tsx
