+++ src/containers/artist-dashboard-page/components/ClaimingModalBody.tsx
