+++ src/containers/artist-dashboard-page/components/DashboardTokenValueSlider.tsx
