--- src/containers/audio-rewards-page/AudioRewardsPage.module.css
@@ -1,30 +1,0 @@
-.pageContainer {
-  user-select: none;
-}
-
-.contentContainer {
-  display: inline-flex;
-  flex-direction: column;
-  width: 1104px;
-  margin: -18px 0px 16px;
-}
-
-.cryptoContentContainer {
-  display: inline-flex;
-  height: 243px;
-  margin-bottom: 32px;
-  width: 100%;
-}
-
-.explainerTile {
-  margin-bottom: 32px;
-}
-
-.balance {
-  margin-right: 24px;
-}
-
-
-.claimTile {
-  margin-right: 24px;
-}
