--- src/containers/audio-rewards-page/components/ClickableAddress.module.css
@@ -1,43 +1,0 @@
-.container {
-  cursor: pointer;
-  transition: all .07s ease-in-out;
-  transform: scale3d(1, 1, 1);
-}
-
-.container:hover {
-  transform: scale3d(1.03, 1.03, 1.03);
-}
-
-.container:active {
-  transform: scale3d(0.99, 0.99, 0.99);
-}
-
-.addressContainer {
-  display: flex;
-  width: 100%;
-  align-items: center;
-}
-
-.address {
-  flex: 1 1;
-  min-width: 0;
-}
-
-.icon {
-  height: 24px;
-  flex: 0 0 24px;
-  margin-left: 12px;
-}
-
-.icon path {
-  fill: var(--static-white);
-}
-
-.toastContainer {
-  position: relative;
-  width: 100%;
-}
-
-.toast {
-  left: 225.5px !important;
-}
