--- src/containers/audio-rewards-page/components/ErrorBody.module.css
@@ -1,33 +1,0 @@
-.container {
-  display: inline-flex;
-  flex-direction: column;
-  justify-content: space-between;
-  align-items: center;
-  padding-top: 32px;
-}
-
-.scrollableMessage {
-  min-height: 132px;
-  height: 100%;
-  max-height: 248px;
-
-  width: 576px;
-  margin: 0px auto 24px; 
-  padding: 24px 20px;
-  color: var(--neutral);
-  background: var(--neutral-light-9);
-  border: 1px solid var(--neutral-light-8);
-  box-sizing: border-box;
-  border-radius: 8px;
-}
-
-.scrollableMessage :global(.simplebar-scrollbar) {
-  width: 6px;
-  border-radius: 4px;
-  opacity: 0.5;
-}
-
-
-.btn {
-  margin-bottom: 24px;
-}
