--- src/containers/audio-rewards-page/components/ExplainerTile.module.css
@@ -1,119 +1,0 @@
-.tileContainer {
-  flex: 1;
-  display: flex;
-  flex-direction: column;
-  justify-content: flex-start;
-  align-items: center;
-  border-radius: 6px;
-  box-shadow: 0px 16px 20px var(--table-shadow);
-  min-height: 210px;
-  width: 100%;
-}
-
-/* Explainer Tile */
-.explainerTile {
-  position: relative;
-  height: 331px;
-  background-color: var(--white);
-  display: inline-flex;
-  flex-wrap: wrap;
-  flex-direction: row;
-  justify-content: flex-start;
-  align-items: flex-start;
-  overflow: hidden;
-  text-align: left;
-}
-
-.whatIsAudioContainer {
-  flex: 1;
-  display: inline-flex;
-  flex-direction: column;
-  align-items: flex-start;
-  padding: 44px 0px;
-}
-
-@media screen and (max-width: 1024px) {
-  .whatIsAudioContainer {
-    padding-right: 32px;
-  }
-}
-
-.whatIsAudio {
-  font-weight: 800;
-  font-size: 40px;
-  line-height: 49px;
-  background: var(--page-header-gradient);
-  -webkit-background-clip: text;
-  -webkit-text-fill-color: transparent;
-  margin: 0px 0px 2px 0px;
-}
-
-.whyCare {
-  font-weight: var(--font-bold);
-  font-size: var(--font-m);
-  line-height: 20px;
-  text-transform: uppercase;
-  color: var(--neutral-light-4);
-  margin-bottom: 18px;
-  text-align: left;
-}
-
-.description {
-  font-weight: var(--font-demi-bold);
-  font-size: 20px;
-  line-height: 160%;
-  color: var(--neutral);
-  margin-bottom: 16px;
-  max-width: 640px;
-}
-
-.learnMore {
-  font-weight: var(--font-bold);
-  font-size: 20px;
-  line-height: 17px;
-  text-align: right;
-  letter-spacing: 0.5px;
-  color: var(--secondary);
-  margin: 10px 0px 8px;
-  cursor: pointer;
-}
-
-.learnMore:hover {
-  text-decoration: underline;
-}
-
-.confused {
-  font-weight: var(--font-demi-bold);
-  font-size: var(--font-m);
-  line-height: 24px;
-  color: var(--neutral);
-  margin-bottom: 28px;
-}
-
-.tokenHero {
-  margin: 66px 64px 66px 102px;
-}
-
-.tokenHero svg, .tokenHero img {
-  width: 200px;
-  height: 200px;
-}
-
-.iconRemove {
-  position: absolute;
-  top: 24px;
-  left: 24px;
-  height: 24px;
-  width: 24px;
-  cursor: pointer;
-}
-.iconRemove:hover {
-  transform: scale(1.10);
-}
-.iconRemove:active {
-  transform: scale(0.95);
-}
-.iconRemove g path {
-  fill: var(--neutral-light-4);
-}
-
