--- src/containers/audio-rewards-page/components/PurpleBox.module.css
@@ -1,39 +1,0 @@
-.container {
-  background: var(--page-header-gradient);
-  box-shadow: 0px 4px 20px -2px rgba(100, 17, 166, 0.2);
-  border-radius: 6px;
-  min-height: 98px;
-  width: 100%;
-  min-width: 593px;
-  padding: 24px 42px;
-  display: inline-flex;
-  flex-direction: column;
-  justify-content: space-between;
-}
-
-.label {
-  font-weight: var(--font-bold);
-  font-size: var(--font-m);
-  line-height: 10px;
-  text-align: center;
-  letter-spacing: 1px;
-  text-transform: uppercase;
-  color: #C675FF;
-  text-shadow: 0px 4px 15px rgba(0, 0, 0, 0.1);
-}
-
-.text {
-  font-weight: bold;
-  font-size: var(--font-l);
-  line-height: 24px;
-  text-align: center;
-  color: var(--static-white);
-}
-
-.compact {
-  line-height: 16px;
-  font-size: var(--font-m);
-  margin-top: 16px;
-  word-wrap: break-word;
-  max-width: 508px;
-}
