--- src/containers/audio-rewards-page/components/TokenHoverTooltip.tsx
@@ -1,29 +1,0 @@
-import React from 'react'
-import Tooltip from 'components/tooltip/Tooltip'
-import { BNWei, formatWei } from 'store/wallet/slice'
-
-type TokenHoverTooltipProps = {
-  children: React.ReactNode
-  balance: BNWei
-  parentMount?: boolean
-}
-
-const TokenHoverTooltip = ({
-  balance,
-  children,
-  parentMount = false
-}: TokenHoverTooltipProps) => {
-  return (
-    <Tooltip
-      text={`${formatWei(balance)} $AUDIO`}
-      disabled={balance.isZero()}
-      placement='top'
-      mouseEnterDelay={0.2}
-      mount={parentMount ? 'parent' : undefined}
-    >
-      {children}
-    </Tooltip>
-  )
-}
-
-export default TokenHoverTooltip
