+++ src/containers/user-badges/TierExplainerModal.tsx
@@ -1,6 +1,6 @@
 import React, { useCallback } from 'react'
 import { Button, ButtonSize, ButtonType, Modal } from '@audius/stems'
-import { Tier } from 'containers/audio-rewards-page/Tiers'
+import { Tier } from 'containers/artist-dashboard-page/Tiers'
 
 import styles from './TierExplainerModal.module.css'
 import { useSelector } from 'utils/reducer'
